Add DailyMetricTriggerSettings. Rework existing code to be reused
